pyproject.toml
[project]
name = "mylac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.10.1",
    "fastmcp>=2.13.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.1.12",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-ollama>=0.3.10",
    "langgraph>=1.0.1",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "pydantic>=2.12.2",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.1",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "qdrant-client[fastembed]>=1.15.1",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
    "starlette>=0.48.0",
    "tqdm>=4.67.1",
    "uuid>=1.30",
]
